requirements.txt
astroid==2.2.5
autopep8==1.5
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.1.7
django-autocomplete-light==3.5.1
django-cors-headers==2.5.2
django-filter==2.1.0
django-rest-swagger==2.2.0
djangorestframework==3.9.2
djangorestframework-camel-case==1.1.2
djangorestframework-jwt==1.11.0
drfdocs==0.0.11
gunicorn==19.9.0
idna==2.8
isort==4.3.16
itypes==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
Pillow==6.0.0
psycopg2==2.8.2
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.3.1
pytz==2018.9
requests==2.21.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.1
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==4.1.2
wrapt==1.11.1
